requirements.txt
streamlit>=1.45.0
pandas>=2.0
plotly>=5.18
spotipy>=2.23
python-dotenv>=1.0
pyarrow>=14.0
